requirements.txt
fastapi
uvicorn
torch
numpy
python-chess
